CFS-121: Split Table Rendering to  its own separate file .
